requirements.txt
Django==2.0
django-bootstrap3==11.0.0
django-simple-captcha==0.5.11
gunicorn==19.9.0
Pillow==6.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
